refactor some wrapper types/interfaces into a separate file we do not want claude to touch
